Move util commands to slashcommand system
