+++ POI.DiscordDotNet/Commands/Modules/ChatCommands/AdminCommandsModule.cs
@@ -1,6 +1,6 @@
 using DSharpPlus.CommandsNext;
 
-namespace POI.DiscordDotNet.Commands.Modules
+namespace POI.DiscordDotNet.Commands.Modules.ChatCommands
 {
 	public class AdminCommandsModule : BaseCommandModule
 	{
