+++ POI.DiscordDotNet/Commands/Modules/ChatCommands/BeatSaberCommandsModule.cs
@@ -1,6 +1,6 @@
 using DSharpPlus.CommandsNext;
 
-namespace POI.DiscordDotNet.Commands.Modules
+namespace POI.DiscordDotNet.Commands.Modules.ChatCommands
 {
 	public class BeatSaberCommandsModule : BaseCommandModule
 	{
