+++ POI.DiscordDotNet/Commands/Modules/SlashCommands/BeatSaberSlashCommandsModule.cs
@@ -1,0 +1,8 @@
+﻿using DSharpPlus.SlashCommands;
+
+namespace POI.DiscordDotNet.Commands.Modules.SlashCommands
+{
+	public class BeatSaberSlashCommandsModule : ApplicationCommandModule
+	{
+	}
+}
