--- POI.DiscordDotNet/Commands/Modules/UtilCommandsModule.cs
@@ -1,8 +1,0 @@
-using DSharpPlus.CommandsNext;
-
-namespace POI.DiscordDotNet.Commands.Modules
-{
-	public class UtilCommandsModule : BaseCommandModule
-	{
-	}
-}
