+++ POI.DiscordDotNet/Commands/Utils/PingCommand.cs
@@ -1,21 +1,18 @@
-﻿using System;
-using System.Threading.Tasks;
-using DSharpPlus.CommandsNext;
-using DSharpPlus.CommandsNext.Attributes;
-using POI.DiscordDotNet.Commands.Modules;
+﻿using System.Threading.Tasks;
+using DSharpPlus.SlashCommands;
+using POI.DiscordDotNet.Commands.Modules.SlashCommands;
 
 namespace POI.DiscordDotNet.Commands.Utils
 {
-    public class PingCommand : UtilCommandsModule
-    {
-	    [Command("ping")]
-        public async Task Handle(CommandContext ctx)
-        {
-            await ctx.Channel
-                .SendMessageAsync("POI!\n" +
-                                  $"WS latency: {ctx.Client.Ping} ms\n" +
-                                  $"Message latency: {(DateTimeOffset.Now - ctx.Message.Timestamp):g}")
-                .ConfigureAwait(false);
-        }
-    }
+	public class PingCommand : UtilSlashCommandsModule
+	{
+		[SlashCommand("ping", "Shows how response I am ^^")]
+		public async Task Handle(InteractionContext ctx)
+		{
+			await ctx
+				.CreateResponseAsync("POI!\n" +
+				                     $"WS latency: {ctx.Client.Ping} ms\n")
+				.ConfigureAwait(false);
+		}
+	}
 }
